Turn off debug mode by default
Debug messages go to stdout, which interferes with my intended use of
that stream in come WALA client code.  I think having DEBUG off by
default is reasonable; clients can turn it on if they need it on.  But
it would be easy for a client to not even be aware that this is an
option, and therefore be in debug mode without ever deciding that the
extra output was something they needed.