Exclude some hand-written HTML test pages from validation
It's unclear whether the original authors of these pages intended them
to be valid or invalid.  Certainly there is merit in testing against
invalid HTML, since the vast majority of real-world HTML is indeed
invalid.  I'm going to assume that any errors in this collection of
test inputs are intentional, and therefore not worth reporting when
running Eclipse HTML validation.