Prefer dynamic dispatch over switch statements on enum values
Eclipse was warning about the lack of a `default` case in the `switch`
statement.  None of the current `enum` values was actually missing
from the `switch`, but if the enum values list were extended in the
future, that would be an easy mistake to make.

Replacing the `switch` with dynamic dispatch to a method lets us
statically enforce the requirement that every enum value must
implement the behavior in some way.