Use the default output folder for `dat`
This fixes an Eclipse Plug-In Development warning that reads "Source
folder 'dat/' does not have the output folder in corresponding output
entry 'output..'."  It is not the default quick fix that Eclipse
suggests, and I am *definitely* not sure that this is the right way to
fix this problem.  (I have nearly zero knowledge of Eclipse plug-in
development.)

That being said, this change does make the warning go away, and `mvn
clean install` without `-DskipTests` still succeeds.