Suppress deprecation warnings about requesting an old AST version
Near as I can tell, the requests for deprecated versions here are
intentional.  The non-deprecated version (AST.JLS9) is the latest and
greatest, but as far as I can tell we really do want the older version
here.

This is similar to 6caecce3e74cec12f4f206bddb39c44875e46ddc, though in
that case JLS8 was the non-reprecated latest version and we were still
asking for JLS3.